refactored instapost into its own component
